+++ lib/rgeo/impl_helper/basic_geometry_methods.rb
@@ -52,7 +52,7 @@ module RGeo
         @factory = obj.factory
       end
 
-      def validate_geometry
+      def prepare_geometry
       end
     end
   end