+++ test/simple_cartesian/calculations_test.rb
@@ -28,6 +28,21 @@ class CartesianCalculationsTest < Minitest::Test # :nodoc:
     @parallel_rising_seg = RGeo::Cartesian::Segment.new(@point6, @point7)
     @steep_rising_seg = RGeo::Cartesian::Segment.new(@point6, @point4)
     @degenerate_seg = RGeo::Cartesian::Segment.new(@point5, @point5)
+
+    # sweepline tests
+    @point9 = @factory.point(0, 0)
+    @point10 = @factory.point(0, 1)
+    @point11 = @factory.point(-0.5, 0.5)
+    @point12 = @factory.point(0.5, 0.5)
+    @point13 = @factory.point(0, 0.5)
+    @point14 = @factory.point(0.5, 0.6)
+    @point15 = @factory.point(0, 0.6)
+
+    @li_seg1 = RGeo::Cartesian::Segment.new(@point9, @point10)
+    @li_seg2 = RGeo::Cartesian::Segment.new(@point11, @point12)
+    @li_seg3 = RGeo::Cartesian::Segment.new(@point12, @point13)
+    @li_seg4 = RGeo::Cartesian::Segment.new(@point14, @point13)
+    @li_seg5 = RGeo::Cartesian::Segment.new(@point14, @point15)
   end
 
   def assert_close_enough(v1, v2)
@@ -90,5 +105,34 @@ class CartesianCalculationsTest < Minitest::Test # :nodoc:
   def test_segment_intersects_endpoints
     assert(@horiz_seg.intersects_segment?(@vert_seg))
     assert(@vert_seg.intersects_segment?(@horiz_seg))
+  end
+
+  def test_segment_intersection_parallel
+    assert_equal(@point5, @long_rising_seg.segment_intersection(@long_rising_seg))
+    assert_nil(@long_rising_seg.segment_intersection(@parallel_rising_seg))
+    assert_equal(@point1, @long_rising_seg.segment_intersection(@short_rising_seg))
+    assert_equal(@point1, @touching_collinear_rising_seg.segment_intersection(@long_rising_seg))
+  end
+
+  def test_segment_intersection_basic
+    int1 = @factory.point(1, 3)
+    assert_equal(int1, @long_rising_seg.segment_intersection(@steep_rising_seg))
+
+    int2 = @factory.point(0, 0.5)
+    assert_equal(int2, @li_seg1.segment_intersection(@li_seg2))
+
+    assert_nil(@short_rising_seg.segment_intersection(@steep_rising_seg))
+  end
+
+  def test_segment_intersection_endpoints
+    assert_equal(@point1, @horiz_seg.segment_intersection(@vert_seg))
+    assert_equal(@point1, @vert_seg.segment_intersection(@horiz_seg))
+  end
+
+  def test_segment_intersection_floating_point_miss
+    poly = @factory.point(0, 0).buffer(1)
+    poly.exterior_ring.segments.each_cons(2) do |seg1, seg2|
+      assert_equal(seg1.e, seg1.segment_intersection(seg2))
+    end
   end
 end